// src/frontend/src/backend.d.ts
import type { Principal } from "@icp-sdk/core/principal";
export interface Some<T> {
    __kind__: "Some";
    value: T;
}
export interface None {
    __kind__: "None";
}
export type Option<T> = Some<T> | None;
export interface TransformationInput {
    context: Uint8Array;
    response: http_request_result;
}
export interface Article {
    title: string;
    body: string;
    published: boolean;
    author: string;
    timestamp: bigint;
}
export interface TransformationOutput {
    status: bigint;
    body: Uint8Array;
    headers: Array<http_header>;
}
export interface http_header {
    value: string;
    name: string;
}
export interface http_request_result {
    status: bigint;
    body: Uint8Array;
    headers: Array<http_header>;
}
export enum UserRole {
    admin = "admin",
    user = "user",
    guest = "guest"
}
export interface backendInterface {
    addToWatchlist(symbol: string): Promise<void>;
    assignCallerUserRole(user: Principal, role: UserRole): Promise<void>;
    clearWatchlist(): Promise<void>;
    createArticle(title: string, body: string, author: string): Promise<void>;
    fetchFinancialNews(symbol: string): Promise<string>;
    fetchMarketSummary(): Promise<string>;
    fetchStockQuote(symbol: string): Promise<string>;
    getCallerUserRole(): Promise<UserRole>;
    getOwnArticles(): Promise<Array<Article>>;
    getPublishedArticles(): Promise<Array<Article>>;
    getWatchlist(): Promise<Array<string>>;
    isCallerAdmin(): Promise<boolean>;
    publishArticle(timestamp: bigint): Promise<void>;
    removeFromWatchlist(symbol: string): Promise<void>;
    transform(input: TransformationInput): Promise<TransformationOutput>;
}